run basedpyright on the client side
