+++ client/package.json
@@ -3,13 +3,15 @@
     "version": "1.0.0",
     "main": "node_modules/expo/AppEntry.js",
     "scripts": {
-        "start": "expo start",
-        "web": "expo start --web",
-        "build:web": "tsc && expo export:web && shx rm -rf ../server/dist/webapp && shx mkdir -p ../server/dist/webapp && shx cp -r ./web-build/* ../server/dist/webapp/"
+        "start": "npm run build:worker && expo start",
+        "web": "npm run start -- --web",
+        "build:worker": "esbuild --bundle worker.ts --outdir=web",
+        "build:web": "npm run build:worker && tsc && expo export:web && shx rm -rf ../server/dist/webapp && shx mkdir -p ../server/dist/webapp && shx cp -r ./web-build/* ../server/dist/webapp/"
     },
     "dependencies": {
         "@ant-design/icons": "^5.2.6",
         "@monaco-editor/react": "^4.6.0",
+        "browser-basedpyright": "^1.14.0-047da5de6154ece025c4d52e72b71ecd9ff074e4",
         "expo": "~49.0.15",
         "expo-status-bar": "~1.6.0",
         "lz-string": "^1.5.0",
@@ -19,12 +21,15 @@
         "react-native": "0.72.6",
         "react-native-popup-menu": "^0.16.1",
         "react-native-web": "~0.19.6",
+        "vscode-jsonrpc": "^8.2.1",
+        "vscode-languageserver-protocol": "^3.17.5",
         "vscode-languageserver-types": "^3.17.5"
     },
     "devDependencies": {
         "@babel/core": "^7.20.0",
         "@expo/webpack-config": "^19.0.0",
         "@types/react": "~18.2.14",
+        "esbuild": "^0.23.0",
         "shx": "^0.3.4",
         "typescript": "^5.1.3"
     },
